* Tasks are properly displayed to the user * Simplfied the date attribute in TaskEntity
